Added category, global, product and shopping basket tests and removed main test file
